refactor(store): Use @endo/(pass-style,patterns,exo) and drop our stale copies
