+++ packages/deploy-script-support/src/depositInvitation.js
@@ -8,7 +8,7 @@ export const makeDepositInvitation = zoeInvitationPurse => {
   /**
    * @param {ERef<Invitation>} invitationP
    * @returns {Promise<InvitationDetails>}
-   * */
+   */
   const depositInvitation = async invitationP => {
     const invitation = await invitationP;
     // Deposit returns the amount deposited
