+++ packages/deploy-script-support/src/writeCoreProposal.js
@@ -1,14 +1,10 @@
 // @ts-check
 import fs from 'fs';
 import { E } from '@endo/far';
+import { deeplyFulfilled } from '@endo/marshal';
 
 import { createBundles } from '@agoric/internal/src/node/createBundles.js';
-import {
-  deeplyFulfilled,
-  defangAndTrim,
-  mergePermits,
-  stringify,
-} from './code-gen.js';
+import { defangAndTrim, mergePermits, stringify } from './code-gen.js';
 import { makeCoreProposalBehavior, permits } from './coreProposalBehavior.js';
 
 export const makeWriteCoreProposal = (
