+++ packages/store/package.json
@@ -32,9 +32,12 @@
   "dependencies": {
     "@agoric/assert": "^0.5.1",
     "@agoric/internal": "^0.2.1",
+    "@endo/exo": "^0.2.1",
     "@endo/far": "^0.2.17",
     "@endo/eventual-send": "^0.17.1",
     "@endo/marshal": "^0.8.4",
+    "@endo/pass-style": "^0.1.2",
+    "@endo/patterns": "^0.2.1",
     "@endo/promise-kit": "^0.2.55",
     "@fast-check/ava": "^1.1.3"
   },
