+++ packages/store/src/stores/store-utils.js
@@ -1,6 +1,31 @@
 import { Far } from '@endo/marshal';
+import { M, matches } from '@endo/patterns';
 
 const { Fail, quote: q } = assert;
+
+/**
+ * Should behave identically to the one in `@endo/patterns`, but
+ * reimplemented for now because `@endo/patterns` forgot to export this one.
+ * This one is simple enough that I prefer a reimplementation to a deep import.
+ * TODO: Undate `@endo/patterns` to export the original, and delete the
+ * reimplementation here.
+ *
+ * @param {Passable} s
+ * @returns {s is CopySet}
+ */
+export const isCopySet = s => matches(s, M.set());
+
+/**
+ * Should behave identically to the one in `@endo/patterns`, but
+ * reimplemented for now because `@endo/patterns` forgot to export this one.
+ * This one is simple enough that I prefer a reimplementation to a deep import.
+ * TODO: Undate `@endo/patterns` to export the original, and delete the
+ * reimplementation here.
+ *
+ * @param {Passable} m
+ * @returns {m is CopyMap}
+ */
+export const isCopyMap = m => matches(m, M.map());
 
 /**
  * @template K,V
