--- packages/store/tools.js
@@ -1,4 +1,0 @@
-export {
-  defendPrototype,
-  defendPrototypeKit,
-} from './src/patterns/exo-tools.js';
