+++ packages/swingset-liveslots/package.json
@@ -20,7 +20,10 @@
   "dependencies": {
     "@agoric/assert": "^0.5.1",
     "@agoric/internal": "^0.2.1",
+    "@endo/exo": "^0.2.1",
     "@endo/nat": "^4.1.26",
+    "@endo/pass-style": "^0.1.2",
+    "@endo/patterns": "^0.2.1",
     "@agoric/store": "^0.8.3",
     "@agoric/vat-data": "^0.4.3",
     "@endo/eventual-send": "^0.17.1",
